fix(binding-command): rename file
