+++ packages/jit/src/index.ts
@@ -1,4 +1,4 @@
 export * from './binding/expression-parser';
-export * from './templating/instruction-compiler';
+export * from './templating/binding-command';
 export * from './templating/template-compiler';
 export * from './configuration';
