+++ packages/jit/src/templating/binding-command.ts
